requirements.txt
pytest==9.0.2
requests==2.32.5
python-dotenv==1.2.1
allure-pytest==2.15.3